Make image capture possible.
The problem with teh canon 70D is that the 'default' storage medium is
the Internal RAM. This means that images taken are not stored on the
camera.

The canon_enable_capture fixes this. Other canon's seem to want some
other setting to do this.
There's also example code (in libgphoto2) that uses half press and full
press to take the photo, that *might* avoid the issue as well.